requirements.txt
python-telegram-bot==20.4
Pillow==10.0.0
opencv-python
pytz==2024.1